pyproject.toml
[tool.poetry]
name = "openai-python"
version = "0.1.0"
description = "An environment for experimenting with the OpenAI API Python library."
authors = ["Christopher Kumm <chrislkumm@protonmail.com>"]
license = "UNLICENSED"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
